Support deny_by_default for bash command tool (#2999)
Problem: Customer wants to allow some commands and deny all other commands.
To do that, today customer needs to add commands to allow list and use a negative
regex in deny list to exclude other commands.

That actually doesn't work. The deny list regex we use in Rust doesn't work for such
negative look around. Given such pattern, the regex will not match anything. This is
bad because customer "deny list" is now not matching anything.

The 1st change is to fallback deny list to "*" when regex compilation fails, instead of
ignoring not matching anything.

The 2nd change is to introduce a boolean flag called "denyByDefault" so anything not in
allowed list can be denied by default (instead of "ask" by default). With this flag,
user can deny by default and explicitly add allowed commands to allowed list.

This flag is OFF by default so it doesn't change default UX.